C++ hackery

+++ src/viper/utils/TimeConstDelay.cpp
@@ -1,5 +1,5 @@
 #include "TimeConstDelay.h"
-#include <cstdlib>
+#include <cmath>
 
 TimeConstDelay::TimeConstDelay() {
     this->offset = 0;
@@ -9,7 +9,7 @@ TimeConstDelay::TimeConstDelay() {
 float TimeConstDelay::ProcessSample(float sample) {
     float val = this->samples[this->offset];
     this->samples[this->offset] = sample;
-    this->offset = (uint32_t) modf((float) this->offset + 1, (float *) &this->sampleCount); // TODO: check if this is correct
+    this->offset = (uint32_t) modff((float) this->offset + 1, (float *) &this->sampleCount); // TODO: check if this is correct
     return val;
 }
